requirements.txt
altair==5.5.0
annotated-types==0.7.0
attrs==25.3.0
azure-core==1.34.0
azure-cosmos==4.9.0
azure-identity==1.23.0
azure-keyvault-secrets==4.9.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.2
cryptography==45.0.4
cycler==0.12.1
dotenv==0.9.9
fonttools==4.58.4
gitdb==4.0.12
gitpython==3.1.44
idna==3.10
isodate==0.7.2
jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
markupsafe==3.0.2
matplotlib==3.10.3
msal==1.32.3
msal-extensions==1.3.1
narwhals==1.42.1
numpy==2.2.6
packaging==24.2
pandas==2.3.0
pillow==11.2.1
protobuf==6.31.1
pyarrow==20.0.0
pycparser==2.22
pydantic==2.11.6
pydantic-core==2.33.2
pydeck==0.9.1
pyjwt==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.25.1
six==1.17.0
smmap==5.0.2
streamlit==1.45.1
tenacity==9.1.2
toml==0.10.2
tornado==6.5.1
typing-extensions==4.14.0
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.4.0
